add rule system change table primary key type because abi error

+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_body_table.hpp
@@ -6,9 +6,9 @@
 class crbody
 {
   public:
-    uint8_t b_body;
+    uint64_t b_body;
 public:
-    uint8_t primary_key() const {return b_body;}
+    uint64_t primary_key() const {return b_body;}
 };
 
 typedef multi_index<N(crbody),crbody> body_data_table;
